ml-service/requirements-macos.txt
# macOS-specific requirements for Apple Silicon (M1/M2/M3/M4)
# Use this instead of requirements.txt on macOS for CoreML acceleration

fastapi==0.135.3
uvicorn==0.42.0
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
pytest==9.0.2
opencv-python-headless==4.13.0.92

# CoreML-optimized ONNX Runtime for Apple Silicon
# Install: pip uninstall onnxruntime && pip install onnxruntime-silicon
onnxruntime-silicon>=1.20,<2; platform_system == "Darwin" and platform_machine == "arm64"

# Fallback to standard onnxruntime for Intel Macs
onnxruntime>=1.20,<2; platform_system == "Darwin" and platform_machine == "x86_64"

insightface>=0.7.3,<1.0
numpy>=1.26

# Orientation detection (CGD model)
torch>=2.0
torchvision>=0.15
timm>=1.0
pytorch-lightning>=2.0
huggingface_hub>=0.20
loguru>=0.7
Pillow>=10.0
